register controller

+++ src/utils/cloudinary.js
@@ -1,30 +1,28 @@
-import {v2 as cloudinary} from "cloudinary";
+import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
-cloudinary.config({ 
-    cloud_name: 'process.env.CLOUDINARY_CLOUD_NAME', 
-    api_key: 'process.env.CLOUDINARY_API_KEY', 
-    api_secret: 'process.env.CLOUDINARY_API_SECRET', 
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
-
-
 const uploadOnCloudinary = async (localFilePath) => {
-    try {
-        if(!localFilePath) return null;
-        // upload the file on the cloudinary
+  try {
+    if (!localFilePath) return null;
+    // upload the file on the cloudinary
     const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
-        }) 
-        // file has been successfully uploaded
-        // console.log(`Uploading file cloudinary`, response.url);
-        return response;
+      resource_type: "auto",
+    });
+    console.log(response);
+    // file has been successfully uploaded
+    fs.existsSync(localFilePath) && fs.unlinkSync(localFilePath);
+    console.log(`Uploading file cloudinary`, response.url);
+    return response;
+  } catch (error) {
+    fs.existsSync(localFilePath) && fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
+    return null;
+  }
+};
 
-    } catch (error) {
-        fs.unlinkSync(localFilePath)  // remove the locally saved temporary file as the upload operation got failed
-        return null;
-    }
-}
-
-
-export {uploadOnCloudinary}
+export { uploadOnCloudinary };
